+++ custom_attributes.go
@@ -23,18 +23,18 @@ import (
 
 type (
 	CustomAttributesServiceInterface interface {
-		ListCustomUserAttributes(user int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error)
-		ListCustomGroupAttributes(group int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error)
-		ListCustomProjectAttributes(project int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error)
-		GetCustomUserAttribute(user int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
-		GetCustomGroupAttribute(group int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
-		GetCustomProjectAttribute(project int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
-		SetCustomUserAttribute(user int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
-		SetCustomGroupAttribute(group int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
-		SetCustomProjectAttribute(project int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
-		DeleteCustomUserAttribute(user int, key string, options ...RequestOptionFunc) (*Response, error)
-		DeleteCustomGroupAttribute(group int, key string, options ...RequestOptionFunc) (*Response, error)
-		DeleteCustomProjectAttribute(project int, key string, options ...RequestOptionFunc) (*Response, error)
+		ListCustomUserAttributes(user int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error)
+		ListCustomGroupAttributes(group int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error)
+		ListCustomProjectAttributes(project int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error)
+		GetCustomUserAttribute(user int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
+		GetCustomGroupAttribute(group int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
+		GetCustomProjectAttribute(project int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
+		SetCustomUserAttribute(user int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
+		SetCustomGroupAttribute(group int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
+		SetCustomProjectAttribute(project int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error)
+		DeleteCustomUserAttribute(user int64, key string, options ...RequestOptionFunc) (*Response, error)
+		DeleteCustomGroupAttribute(group int64, key string, options ...RequestOptionFunc) (*Response, error)
+		DeleteCustomProjectAttribute(project int64, key string, options ...RequestOptionFunc) (*Response, error)
 	}
 
 	// CustomAttributesService handles communication with the group, project and
@@ -60,7 +60,7 @@ type CustomAttribute struct {
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#list-custom-attributes
-func (s *CustomAttributesService) ListCustomUserAttributes(user int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) ListCustomUserAttributes(user int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
 	return s.listCustomAttributes("users", user, options...)
 }
 
@@ -68,7 +68,7 @@ func (s *CustomAttributesService) ListCustomUserAttributes(user int, options ...
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#list-custom-attributes
-func (s *CustomAttributesService) ListCustomGroupAttributes(group int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) ListCustomGroupAttributes(group int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
 	return s.listCustomAttributes("groups", group, options...)
 }
 
@@ -76,11 +76,11 @@ func (s *CustomAttributesService) ListCustomGroupAttributes(group int, options .
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#list-custom-attributes
-func (s *CustomAttributesService) ListCustomProjectAttributes(project int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) ListCustomProjectAttributes(project int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
 	return s.listCustomAttributes("projects", project, options...)
 }
 
-func (s *CustomAttributesService) listCustomAttributes(resource string, id int, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) listCustomAttributes(resource string, id int64, options ...RequestOptionFunc) ([]*CustomAttribute, *Response, error) {
 	u := fmt.Sprintf("%s/%d/custom_attributes", resource, id)
 	req, err := s.client.NewRequest(http.MethodGet, u, nil, options)
 	if err != nil {
@@ -99,7 +99,7 @@ func (s *CustomAttributesService) listCustomAttributes(resource string, id int, 
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#single-custom-attribute
-func (s *CustomAttributesService) GetCustomUserAttribute(user int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) GetCustomUserAttribute(user int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	return s.getCustomAttribute("users", user, key, options...)
 }
 
@@ -107,7 +107,7 @@ func (s *CustomAttributesService) GetCustomUserAttribute(user int, key string, o
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#single-custom-attribute
-func (s *CustomAttributesService) GetCustomGroupAttribute(group int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) GetCustomGroupAttribute(group int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	return s.getCustomAttribute("groups", group, key, options...)
 }
 
@@ -115,11 +115,11 @@ func (s *CustomAttributesService) GetCustomGroupAttribute(group int, key string,
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#single-custom-attribute
-func (s *CustomAttributesService) GetCustomProjectAttribute(project int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) GetCustomProjectAttribute(project int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	return s.getCustomAttribute("projects", project, key, options...)
 }
 
-func (s *CustomAttributesService) getCustomAttribute(resource string, id int, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) getCustomAttribute(resource string, id int64, key string, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	u := fmt.Sprintf("%s/%d/custom_attributes/%s", resource, id, key)
 	req, err := s.client.NewRequest(http.MethodGet, u, nil, options)
 	if err != nil {
@@ -138,7 +138,7 @@ func (s *CustomAttributesService) getCustomAttribute(resource string, id int, ke
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#set-custom-attribute
-func (s *CustomAttributesService) SetCustomUserAttribute(user int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) SetCustomUserAttribute(user int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	return s.setCustomAttribute("users", user, c, options...)
 }
 
@@ -146,7 +146,7 @@ func (s *CustomAttributesService) SetCustomUserAttribute(user int, c CustomAttri
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#set-custom-attribute
-func (s *CustomAttributesService) SetCustomGroupAttribute(group int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) SetCustomGroupAttribute(group int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	return s.setCustomAttribute("groups", group, c, options...)
 }
 
@@ -154,11 +154,11 @@ func (s *CustomAttributesService) SetCustomGroupAttribute(group int, c CustomAtt
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#set-custom-attribute
-func (s *CustomAttributesService) SetCustomProjectAttribute(project int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) SetCustomProjectAttribute(project int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	return s.setCustomAttribute("projects", project, c, options...)
 }
 
-func (s *CustomAttributesService) setCustomAttribute(resource string, id int, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
+func (s *CustomAttributesService) setCustomAttribute(resource string, id int64, c CustomAttribute, options ...RequestOptionFunc) (*CustomAttribute, *Response, error) {
 	u := fmt.Sprintf("%s/%d/custom_attributes/%s", resource, id, c.Key)
 	req, err := s.client.NewRequest(http.MethodPut, u, c, options)
 	if err != nil {
@@ -177,7 +177,7 @@ func (s *CustomAttributesService) setCustomAttribute(resource string, id int, c 
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#delete-custom-attribute
-func (s *CustomAttributesService) DeleteCustomUserAttribute(user int, key string, options ...RequestOptionFunc) (*Response, error) {
+func (s *CustomAttributesService) DeleteCustomUserAttribute(user int64, key string, options ...RequestOptionFunc) (*Response, error) {
 	return s.deleteCustomAttribute("users", user, key, options...)
 }
 
@@ -185,7 +185,7 @@ func (s *CustomAttributesService) DeleteCustomUserAttribute(user int, key string
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#delete-custom-attribute
-func (s *CustomAttributesService) DeleteCustomGroupAttribute(group int, key string, options ...RequestOptionFunc) (*Response, error) {
+func (s *CustomAttributesService) DeleteCustomGroupAttribute(group int64, key string, options ...RequestOptionFunc) (*Response, error) {
 	return s.deleteCustomAttribute("groups", group, key, options...)
 }
 
@@ -193,11 +193,11 @@ func (s *CustomAttributesService) DeleteCustomGroupAttribute(group int, key stri
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/custom_attributes/#delete-custom-attribute
-func (s *CustomAttributesService) DeleteCustomProjectAttribute(project int, key string, options ...RequestOptionFunc) (*Response, error) {
+func (s *CustomAttributesService) DeleteCustomProjectAttribute(project int64, key string, options ...RequestOptionFunc) (*Response, error) {
 	return s.deleteCustomAttribute("projects", project, key, options...)
 }
 
-func (s *CustomAttributesService) deleteCustomAttribute(resource string, id int, key string, options ...RequestOptionFunc) (*Response, error) {
+func (s *CustomAttributesService) deleteCustomAttribute(resource string, id int64, key string, options ...RequestOptionFunc) (*Response, error) {
 	u := fmt.Sprintf("%s/%d/custom_attributes/%s", resource, id, key)
 	req, err := s.client.NewRequest(http.MethodDelete, u, nil, options)
 	if err != nil {
